Fix par spacing issue, add preamble page generation
This commit fixes some spacing issues between paragraphs and lines. By increasing the spacing, it matches much more closely to other templates, but also improves readability. The only position where this rule is "violated" is when the thesis title is being rendered, but that's because it would look weird otherwise.

The code that generates the title page has also been split off into its own function (`_make_titlepage()`), which is called from within the thesis function. The title page function is prefixed with an underscore to indicate that it shouldn't be called by the user.

The preamble pages (such as the acknowledgements and abstract) get their own function (`_make_preamble_page()`), which makes it easy to reuse the same code to generate identical looking pages. An example of how it is used is provided with the acknowledgements page, where it accepts a global variable `ack_body` as its content. `ack_body` is set by the `acknowledgements()` function, which is intended to be used by the user as a nice way to generate the page, abstracting the numbering, headers & footers and more.