Modify bibliography input to require a call to `bibliography()` by the user
Since this project is intended to be used as a template, it's important that the user can input their bibliography. However, Typst only resolves paths within either the project directory or the template directory.

In the previous version, simply passing the path to the bibliography file was enough as the template source code was stored within the test project that I was using, but since seperating them apart and making this a proper template, it ran into "file not found" errors since it was resolving in the wrong area.

This commit makes it so that the user has to pass a `bibliography()` call into `bib`, so that it can be styled and appended properly.