Fix figure counting
This commit ensures that the figures are being properly counted so that they are local to their chapter, and that the correct text is being generated (i.e. 1.1 for chapter 1, figure 1, A.2 for appendix A, figure 2 etc...)

This is done by clearing the counters at the start of every chapter (H1) via a call to `_clear_counters()`. There is a check before clearing the counters which updates some global variables to ensure that even if the counter has been cleared and not updated since, that if there was content added, it gets properly shown in the preamble/listing of {tables, equations, listings}.

Closes #9
Closes #6