Add modified citation style sheet (CSL)
This commit adds a variant of the `harvard-cite-them-right` 12th edition CSL sheet. It has been modified to change the brackets for the "Available at" section from the round ones "()" to the square ones "[]", and to make sure that the "accessed" text is all lowercase, as per the university referencing guidebook.